+++ CHANGELOG.md
@@ -4,9 +4,14 @@ Changelog
 [0.2.0] (unreleased)
 --------------------
 
+### All
+
+- Everything marked with `@Deprecated(forRemoval=true)` has been removed
+
 ### Core
 
 - Features with a protected constructor are no longer rejected for mounting
+- Extending a feature without `@Import` is now an error
 
 ### HTTP
 
@@ -62,7 +67,7 @@ Sorry, no changelog (yet).
 
 
 
-[0.2.0]: https://github.com/Abnaxos/meldioc/compare/release/0.1.6...develop/next
+[0.2.0]: https://github.com/Abnaxos/meldioc/compare/release/0.1.7...develop/next
 [0.1.7]: https://github.com/Abnaxos/meldioc/compare/release/0.1.6...release/0.1.7
 [0.1.6]: https://github.com/Abnaxos/meldioc/compare/release/0.1.5...release/0.1.6
 [0.1.5]: https://github.com/Abnaxos/meldioc/tags